Move comments into their own file
